requirements.txt
streamlit
bardapi
google.generativeai
streamlit_chat
extra_streamlit_components
# python-dotenv
# langchain
# PyPDF2
# chromadb
# faiss-cpu
# langchain_google_genai